chore(ci): fix file-size cap + typos on the merged branch
msg_entry.rs (508) and hint_injector.rs (510) crossed the 500-line cap
after the main merge — test modules moved to sibling *_tests.rs files
per repo idiom (path-attr modules, zero test changes beyond one extra
super hop). The typos failures were the deliberate close-tag ladder
prefixes ('</paramete' etc.) asserted verbatim in the tool-grammar
EBNF tests — whitelisted in _typos.toml with rationale.
