+++ .github/workflows/file-size-cap.yml
@@ -171,6 +171,12 @@ jobs:
             "crates/spark-server/src/scheduler/emit_step.rs"
             "crates/spark-server/src/scheduler/helpers.rs"
             "crates/spark-server/src/tool_parser/validation.rs"
+            # 510 LoC — prefill-b step: grew by the #283 param_close_pending
+            # field addition. Split tracked as follow-up.
+            "crates/spark-server/src/scheduler/prefill_b_step.rs"
+            # 516 LoC — verify pipeline helpers: grew by the #283
+            # mtp_verify_sample_enabled + dflash helpers. Split tracked as follow-up.
+            "crates/spark-server/src/scheduler/verify_pipeline_helper.rs"
             # ── PR #169 (perf/qwen36-dense-ffn-tc-prefill) carry-over,
             # grown by PR #223 (perf/qwen36-27b-prefill-nvfp4) ──
             # 1507 LoC — dense-FFN forward: multi-arm quantized-FFN dispatch
