+++ crates/spark-server/src/scheduler/prefill_b_step.rs
@@ -169,6 +169,8 @@ pub fn prefill_request(
             min_tokens: req_min_tokens,
             eos_tokens: eos_tokens.to_vec(),
             finished: true,
+            guard_stop: None,
+            param_close_pending: 0,
             sink,
             cancel_flag: cancel_flag.clone(),
             temperature,
